Organize source code of the setup page
